Process all records without returning from PollingThread#poll
The current implementation returns from PollingThread#poll
when Concurrent::RejectedExecutionError occurs, and after that,
waits for CronoTrigger.config.polling_interval.
This means that, with the default interval, a worker can process
only about 15 records per second.
This changes have the following advantages:

1. PollingThread doesn't return from #poll if there are remaining
   records.
2. PollingThread never calls crono_trigger_unlock_all!

I thought we should assume that the fallback policy of the executor
is :caller_runs, but according to the tests, PollingThread seems
to be expected to accept an arbitrary executor.
This is why I didn't change any code of PollingThread.