Change the type of crono_trigger_workers.polling_model_names to text
The polling_model_names column is a string, but it can be a longer
string than 255 characters, which is the upper limit of the string
length of `varchar(255)` in MySQL, and result in an error.
Therefore, it would be better to change the column type to text.

The column contains a comma-separated list of models, so the length of
the string can reach the limit when the number of models increases.
It should not impact performance since the column is not indexed.